Move ProjectSdkImplicitImport tests to own file
